pyproject.toml
[tool.poetry]
name = "new-passion-fruit"
version = "0.1.0"
description = ""
authors = ["ducknam3650 <ducknam3650@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.120.0"
uvicorn = "^0.38.0"
ruff = "^0.14.2"
tortoise-orm = "^0.25.1"
asyncpg = "^0.30.0"
python-dotenv = "^1.0.1"
selenium = "^4.38.0"
beautifulsoup4 = "^4.14.2"
webdriver-manager = "^4.0.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
email-validator = "^2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
